requirements.txt
customtkinter
asyncio
aiohttp
Flask-socketio